app/requirements.txt

plotly==5.10.0
antlr4-python3-runtime==4.7.2
black==22.10.0
Brotli==1.0.9
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
dash==2.6.2
dash-bootstrap-components==1.2.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-iconify==0.1.2
dash-mantine-components==0.10.2
dash-table==5.0.0
dataclasses-json==0.5.7
Flask==2.2.2
Flask-Compress==1.13
flatbuffers==22.9.24
funcsigs==1.0.2
future==0.18.2
google-api-core==2.10.2
google-auth==2.13.0
google-cloud-vision==3.1.4
googleapis-common-protos==1.56.4
grpcio==1.50.0
grpcio-status==1.50.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
keras==2.10.0
latex2sympy2==1.7.9
libclang==14.0.6
MarkupSafe==2.1.1
marshmallow==3.18.0
marshmallow-enum==1.5.1
mpmath==1.2.1
mypy-extensions==0.4.3
numexpr==2.8.4
numpy==1.23.4
opencv-python==4.6.0.66
orjson==3.8.0
packaging==21.3
pandas==1.5.0
pathspec==0.10.1
pdf2image==1.16.0
Pillow==9.2.0
platformdirs==2.5.2
plotly==5.10.0
proto-plus==1.22.1
protobuf==4.21.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
pytesseract==0.3.10
python-dateutil==2.8.2
pytz==2022.5
reportlab==3.6.12
requests==2.28.1
rsa==4.9
shutilwhich==1.1.0
six==1.16.0
sympy==1.11.1
tempdir==0.7.1
tenacity==8.1.0
tensorboard-plugin-wit==1.8.1
tomli==2.0.1
tqdm==4.64.1
typing-inspect==0.8.0
typing_extensions==4.4.0
urllib3==1.26.12
Werkzeug==2.2.2
wget==3.2
win32-setctime==1.1.0
wrapt==1.14.1
